Bugfix: `getPredictedTime` with no comparison
When using `getAll()`, `getPredictedTime()` will not be passed an
argument for the comparison, which causes it to send the message as
"getpredictedtime undefined", throwing an error in LiveSplit.Server.
This commit adds a catch for an undefined argument in
`getPredictedTime`. It will still throw an error in the current version
of LiveSplit.Server since it currently requires a parameter, but I have
a PR open that fixes this behavior in that component.

Additional changes:
 - Removed some superfluous conditions in `if` statements