Use a nonce to guarantee message responses
LiveSplit.Server is not guaranteed to respond to messages in the same
order they are received. This change uses the "nonce" field in the new
JSON messaging standard to provide a unique identifier against which to
match incoming and outgoing messages.

Currently, `this._openRequests` is not actually used for anything. But
it does provide a useful method for investigating errors or commands
that never receive a response. It also may be desired to change the
listening behavior of this library to use a single listener that looks
up the appropriate Promise resolver in `this._openRequests` instead of
the current behavior -- subscribing a new listener for every message.